+++ lib/ductwork/processes/thread_supervisor.rb
@@ -91,7 +91,7 @@ module Ductwork
         workers.each(&:stop)
         await_threads_graceful_shutdown
         kill_remaining_threads
-        delete_process!
+        destroy_process_record!
         run_hooks_for(:stop)
       end
 
@@ -152,12 +152,9 @@ module Ductwork
         end
       end
 
-      def delete_process!
+      def destroy_process_record!
         Ductwork.wrap_with_app_executor do
-          Ductwork::Process.find_by(
-            pid: ::Process.pid,
-            machine_identifier: Ductwork::MachineIdentifier.fetch
-          )&.delete
+          Ductwork::Process.destroy_current!
         end
       end